Extract header className logic into a separate hook
